requirements.txt
python-telegram-bot==13.12
environs==9.5.0
requests==2.27.1
google-cloud-dialogflow==2.19.1
vk-api==11.9.9

